move vs-focus-trap
