+++ packages/vlossom/src/components/vs-checkbox/VsCheckboxSet.vue
@@ -59,7 +59,7 @@
 import { computed, defineComponent, PropType, Ref, ref, toRefs } from 'vue';
 import { useColorScheme, useStyleSet, useInput, useInputOption } from '@/composables';
 import { getInputProps, getInputOptionProps, getResponsiveProps } from '@/models';
-import { VsComponent, VsNode, type ColorScheme } from '@/declaration';
+import { VsComponent, type ColorScheme } from '@/declaration';
 import { utils } from '@/utils';
 import VsInputWrapper from '@/components/vs-input-wrapper/VsInputWrapper.vue';
 import VsCheckboxNode from '@/components/vs-checkbox/VsCheckboxNode.vue';
@@ -132,7 +132,7 @@ export default defineComponent({
             styleSet,
         );
         const { plainStyleSet: checkboxNodeStyleSet } = useStyleSet<VsCheckboxNodeStyleSet>(
-            VsNode.VsCheckboxNode,
+            VsComponent.VsCheckboxNode,
             styleSet,
             checkboxSetStyleSet,
         );
