+++ packages/vlossom/src/nodes/index.ts
@@ -3,8 +3,6 @@ export { default as VsConfirmation } from './vs-confirm/VsConfirmation.vue';
 
 export { default as VsContentRenderer } from './vs-content-renderer/VsContentRenderer.vue';
 
-export { default as VsFocusTrap } from './vs-focus-trap/VsFocusTrap.vue';
-
 export type { VsModalNodeStyleSet, VsModalOptions } from './vs-modal-node/types';
 export { default as VsModalNode } from './vs-modal-node/VsModalNode.vue';
 export { default as VsModalView } from './vs-modal-node/VsModalView.vue';
